+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Note the first digit of every adapter version corresponds to the major version o
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
 ### 4.6.2.0.0
+- Fixed partner SDK version to match the value reported by the SDK.
 - This version of the adapters has been certified with HyprMX 6.2.0.
 
 ### 4.6.0.0.0