+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Note the first digit of every adapter version corresponds to the major version o
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
 ### 4.6.2.0.1
+- Pass COPPA setting HyprMX at startup.
+- Add logging for changes to privacy consent status.
+- Fix missing callback when banner ads are clicked.
 - This version of the adapters has been certified with HyprMX 6.2.0.
 
 ### 4.6.2.0.0